Model development and cors filter added

+++ AngelicaRestService/src/main/java/com/angelica/dao/ProductDao.java
@@ -6,6 +6,6 @@ import com.angelica.model.Product;
 
 public interface ProductDao {
 	
-	List<Product> getAllowedProductsByCustomer(Integer customerId);
+	List<Product> getAllowedProductsByCustomer(Long customerId);
 
 }
